+++ src/views/tool/generator/index.vue
@@ -11,6 +11,7 @@
           <el-input
             v-model="queryParams.tableName"
             placeholder="请输入表名称"
+            clearable
             @keyup.enter="handleList"
           />
         </el-form-item>
@@ -18,6 +19,7 @@
           <el-input
             v-model="queryParams.tableComment"
             placeholder="请输入表描述"
+            clearable
             @keyup.enter="handleList"
           />
         </el-form-item>
@@ -55,10 +57,18 @@
             @click="handleImportTable"
           >导入</el-button>
           <el-button
+            type="success"
+            :icon="Edit"
+            :disabled="genTableIds.length !== 1"
+            v-hasPermission="'tool:generator:edit'"
+            @click="handleEdit"
+          >修改</el-button>
+          <el-button
             type="danger"
             :icon="Delete"
+            :disabled="genTableIds.length === 0"
             v-hasPermission="'tool:generator:delete'"
-            @click="handleDelete"
+            @click="handleDeletes"
           >删除</el-button>
         </el-row>
       </template>
@@ -67,10 +77,6 @@
       <el-table
         v-loading="loading"
         :data="genTableList"
-        row-key="tableId"
-        lazy
-        default-expand-all
-        :tree-props="{ children: 'children', hasChildren: 'hasChildren' }"
         @selection-change="handleSelectionChange"
       >
         <el-table-column type="selection" width="55" />
@@ -82,17 +88,17 @@
         <el-table-column
           prop="tableName"
           label="表名称"
-          align="center"
+          show-overflow-tooltip
         />
         <el-table-column
           prop="tableComment"
           label="表描述"
-          align="center"
+          show-overflow-tooltip
         />
         <el-table-column
           prop="className"
           label="类名称"
-          align="center"
+          show-overflow-tooltip
         />
         <el-table-column
           prop="createTime"
@@ -108,8 +114,8 @@
         />
         <el-table-column
           label="操作"
-          width="250"
           align="center"
+          width="250"
         >
           <template #default="scope">
             <el-button
@@ -166,9 +172,9 @@
 
 <script lang="ts" setup>
 import { getCurrentInstance, onMounted, reactive, ref, toRefs } from 'vue'
-import { ElMessage, FormInstance } from 'element-plus'
+import { FormInstance } from 'element-plus'
 import { Edit, Delete, Search, Refresh, Upload, View, Download } from '@element-plus/icons-vue'
-import { listGenTable, editGenTable, deleteGenTable, downloadCode, downloadCodes } from '@/api/tool/generator'
+import { listGenTable, deleteGenTable, deleteGenTables, downloadCode, downloadCodes } from '@/api/tool/generator'
 import { GenTable, GenTableForm, GenTableQuery } from '@/api/tool/generator/types'
 import { zip } from '@/utils/download'
 import ImportTable from './importTable.vue'
@@ -196,12 +202,6 @@ const state = reactive({
     pageNum: 1,
     pageSize: 10
   } as GenTableQuery,
-  // 对话框
-  genTableDialog: {
-    title: '',
-    type: '',
-    visible: false
-  } as Dialog,
   // 表单
   genTableForm: {} as GenTableForm
 })
@@ -211,15 +211,12 @@ const {
   genTableIds,
   total,
   genTableList,
-  queryParams,
-  genTableDialog,
-  genTableForm
+  queryParams
 } = toRefs(state)
 
 const importTableRef = ref<InstanceType<typeof ImportTable>>()
 const previewCodeRef = ref<InstanceType<typeof PreviewCode>>()
 const editTableRef = ref<InstanceType<typeof EditTable>>()
-const genTableRuleFormRef = ref<FormInstance>()
 const genTableQueryFormRef = ref<FormInstance>()
 
 // 查询代码生成列表
@@ -244,14 +241,34 @@ function handlePreviewCode(row: any) {
 
 // 修改代码生成信息
 function handleEdit(row: any) {
-  editTableRef.value?.handleShow(row.tableId)
+  let genTableId = state.genTableId
+  if (row.tableId) {
+    genTableId = row.tableId
+  }
+  editTableRef.value?.handleShow(genTableId)
 }
 
 // 删除代码生成信息
 function handleDelete(row: any) {
-  deleteGenTable(row.tableId).then(() => {
-    ElMessage.success("删除代码生成信息成功")
-    handleList()
+  proxy.$confirm('确认要删除"' + row.tableNames + '"代码生成信息吗？', '提示', {
+    type: 'warning'
+  }).then(() => {
+    deleteGenTable(row.tableId).then(() => {
+      proxy.$message.success("删除代码生成信息成功")
+      handleList()
+    })
+  })
+}
+
+// 批量删除代码生成信息
+function handleDeletes() {
+  proxy.$confirm('确认要删除"' + state.tableNames.toString() + '"代码生成信息吗？', '提示', {
+    type: 'warning'
+  }).then(() => {
+    deleteGenTables(state.genTableIds).then(() => {
+      proxy.$message.success("删除代码生成信息成功")
+      handleList()
+    })
   })
 }
 
@@ -276,38 +293,13 @@ function resetQuery() {
   handleList()
 }
 
-// 关闭对话框
-function closeGenTableDialog() {
-  state.genTableDialog.visible = false
-  genTableRuleFormRef.value?.clearValidate()
-  genTableRuleFormRef.value?.resetFields()
-}
-
 // 多选框
 function handleSelectionChange(selection: any) {
-  state.genTableIds = selection.map((item: any) => item.genTableId)
+  state.genTableIds = selection.map((item: any) => item.tableId)
   state.tableNames = selection.map((item: any) => item.tableName)
   if (selection.length === 1) {
     state.genTableId = genTableIds.value[0]
   }
-}
-
-// 提交表单
-const submitForm = async (formEl: FormInstance | undefined) => {
-  if (!formEl) return
-  await formEl.validate((valid, fields) => {
-    if (valid) {
-      if (genTableDialog.value.type == 'edit') {
-        editGenTable(state.genTableForm).then(() => {
-          ElMessage.success("修改代码生成信息成功")
-          closeGenTableDialog()
-          handleList()
-        })
-      }
-    } else {
-      console.log('error submit!', fields)
-    }
-  })
 }
 
 onMounted(() => {
